ColdFire: DCR is a 16-bit register
git-svn-id: svn://svn.rockbox.org/rockbox/trunk@6604 a1c6a512-1295-4272-9138-f99709370657

+++ firmware/export/mcf5249.h
@@ -44,7 +44,7 @@
 #define CSMR3 (*(volatile unsigned long *)(MBAR + 0x0a8))
 #define CSCR3 (*(volatile unsigned long *)(MBAR + 0x0ac))
 
-#define DCR   (*(volatile unsigned long *)(MBAR + 0x100))
+#define DCR   (*(volatile unsigned short *)(MBAR + 0x100))
 #define DACR0 (*(volatile unsigned long *)(MBAR + 0x108))
 #define DMR0  (*(volatile unsigned long *)(MBAR + 0x10c))
 #define DACR1 (*(volatile unsigned long *)(MBAR + 0x110))
